Split serializers into their respective view files
